+++ pages/dashboards/ocean-watch/country/[iso].jsx
@@ -13,6 +13,7 @@ import {
 import { dehydrate } from 'react-query/hydration';
 import dynamic from 'next/dynamic';
 import { useRouter } from 'next/router';
+import Link from 'next/link';
 import {
   useSelector,
 } from 'react-redux';
@@ -32,6 +33,7 @@ import NumericCardIndicator from 'components/card-indicator-set/numeric-card-ind
 import MapWidget from 'components/widgets/types/map';
 import SwipeMapWidget from 'components/widgets/types/map-swipe';
 import ChartWidget from 'components/widgets/types/chart';
+import Banner from 'components/app/common/Banner';
 
 // hooks
 import {
@@ -344,6 +346,125 @@ export default function OceanWatchCountryProfilePage({
             </div>
           </section>
         ))}
+        <div className="l-section">
+          <div className="row">
+            <div className="column small-12">
+              <Banner
+                useDim
+                className="-text-center"
+                bgImage="/static/images/pages/app/banner-coral.jpg"
+              >
+                <p className="-claim">
+                  Check the Coral Reefs dashboard
+                </p>
+                <a
+                  className="c-button -alt -primary"
+                  href="https://resourcewatch.org/dashboards/coral-reefs"
+                >
+                  Coral Reefs
+                </a>
+              </Banner>
+            </div>
+            <div
+              className="column small-6"
+              style={{
+                margin: '20px 0 0',
+              }}
+            >
+              <Banner
+                bgImage="/static/images/homepage/home-data-bg1.png"
+                styles={{
+                  padding: 40,
+                }}
+              >
+                <div
+                  style={{
+                    display: 'flex',
+                    flexDirection: 'column',
+                    alignItems: 'flex-start',
+                  }}
+                >
+                  <div>
+                    <p className="-claim">
+                      Did you miss something?
+                    </p>
+                    <p>
+                      Know of a data set that you&apos;d like to see on Resource Watch or have
+                      a specific area of interest you&apos;d like us to cover?
+                    </p>
+                  </div>
+                  <Link href="/get-involved/contribute-data">
+                    <a className="c-button -alt -primary">
+                      Request data
+                    </a>
+                  </Link>
+                </div>
+              </Banner>
+            </div>
+            <div
+              className="column small-6"
+              style={{
+                margin: '20px 0 0',
+              }}
+            >
+              <Banner
+                bgImage="/static/images/backgrounds/jellyfish.jpg"
+                styles={{
+                  padding: 40,
+                  height: '100%',
+                }}
+              >
+                <div
+                  style={{
+                    display: 'flex',
+                    flexDirection: 'column',
+                    alignItems: 'flex-start',
+                    height: '100%',
+                    justifyContent: 'space-between',
+                  }}
+                >
+                  <div>
+                    <p className="-claim">
+                      Suggest a story
+                    </p>
+                    <p>
+                      Have you uncovered an interesting connection in the data
+                      we should tell a story about?
+                    </p>
+                  </div>
+                  <Link href="/get-involved/suggest-a-story">
+                    <a className="c-button -alt -primary">
+                      Suggest a story
+                    </a>
+                  </Link>
+                </div>
+              </Banner>
+            </div>
+            <div
+              className="column small-12"
+              style={{
+                margin: '20px 0 0',
+              }}
+            >
+              <Banner
+                useDim
+                className="-text-center"
+                bgImage="/static/images/pages/app/banner-ocean-watch.jpg"
+              >
+                <p className="-claim">
+                  Check out the Ocean Watch data
+                  <br />
+                  on the Explore page
+                </p>
+                <Link href="/data/explore?section=All data&topics=[&quot;ocean&quot;]">
+                  <a className="c-button -alt -primary">
+                    Go to explore
+                  </a>
+                </Link>
+              </Banner>
+            </div>
+          </div>
+        </div>
       </div>
       {(!!widgetToShare) && (
         <WidgetShareModal